requirements.txt
flask
WeasyPrint
python-dotenv
gunicorn
